Fixed some problem in ZookeeperDoubleBarrier.
Remaining issues:
When the last thread (named A) reaches barrier, creates it's node,
it may notify another thread's (named B) watcher, and cause thread B
to enter the barrier (before thread A). It open a gate for thread B
to leave the barrier, and remove it's node. This may happen before
thread A's watcher is called. So thread A may miss the occasion that
the count of node satisfy the requirement, beause zookeeper won't
send notification on every change.